Splitting up session files
