+++ ParraCore/Types/Sessions/ParraSessionEventType.swift
@@ -1,0 +1,13 @@
+//
+//  ParraSessionEventType.swift
+//  ParraCore
+//
+//  Created by Mick MacCallum on 12/29/22.
+//  Copyright © 2022 Parra, Inc. All rights reserved.
+//
+
+import Foundation
+
+internal enum ParraSessionEventType: String {
+    case impression, log
+}
